# Shadow cascade pass optimization by rejecting more meshes based on frustum intersection between light and camera.