@invertednormal 2023-06-01
If I understand it right, I'd think this should be an optional optimization since it can ignore shadow casters that are outside the view's frustum that still result in a shadow that the view frustum can see. Only draws outside the view frustum opposite the direction of the light are actually safe to cull.