@rgba16f 2023-06-01
> > If I understand it right, I'd think this should be an optional optimization since it can ignore shadow casters that are outside the view's frustum that still result in a shadow that the view frustum can see. Only draws outside the view frustum opposite the direction of the light are actually safe to cull.
> 
> I am rotating the Octree OBB towards light direction and then extruding the Octee's OBB by 36% (bigger OBBs will extend further) + another 20 meters in order to catch nodes that are outside the frustum but still cast shadow into the frustum. It is not ideal but it should be good for most cases and keeps the cpu cost down. In theory, I could calculate light direction vector to camera frustum intersection and extrude to that distance but then the cpu cost will become significant enough that this optimization is not worth.
> 
> A possible direction I was considering was to have the extrusion direction set to -1 as default so the optimization is disabled by default but if a content developer sets this cvar to a positive number, it triggers this optimization and it forces the developer to set the correct amount of extrusion for their content. So if you have a very large scene with massive buildings the content developers can set the extrusion cvar to 50m in their project. I am curious to know your thoughts @invertednormal.

I think @invertednormal is thinking of cases where a directional light (or any other high intensity light) is low to the horizon and will cast shadows across most of the level. The optimization will make it faster at the cost of having incorrect shadows in some cases.