@mbalfour-amzn 2023-06-01
> > > > If I understand it right, I'd think this should be an optional optimization since it can ignore shadow casters that are outside the view's frustum that still result in a shadow that the view frustum can see. Only draws outside the view frustum opposite the direction of the light are actually safe to cull.
> > > 
> > > 
> > > I am rotating the Octree OBB towards light direction and then extruding the Octee's OBB by 36% (bigger OBBs will extend further) + another 20 meters in order to catch nodes that are outside the frustum but still cast shadow into the frustum. It is not ideal but it should be good for most cases and keeps the cpu cost down. In theory, I could calculate light direction vector to camera frustum intersection and extrude to that distance but then the cpu cost will become significant enough that this optimization is not worth.
> > > A possible direction I was considering was to have the extrusion direction set to -1 as default so the optimization is disabled by default but if a content developer sets this car to a positive number, it triggers this optimization and it forces the developer to set the correct amount of extrusion for their content. So if you have a very large scene with massive buildings the content developers can set the extrusion cvar to 50m in their project. I am curious to know your thoughts @invertednormal.
> > 
> > 
> > Should this optimization be disabled for any level that uses terrain?
> 
> No. Its not about using terrain or not. Its more about how much to extrude the OBB to catch some weird edge cases for your content. You can get these edge cases with or without terrain.

To rephrase the question, terrain is a use case that very often has tall features that can appear very far behind the camera. If this optimization is turned on by default, won't this be extremely likely to cause artifacts in most standard terrain use cases? 

Assuming I'm understanding the ramifications correctly, I'm basically just +1ing the idea of having this optimization off by default, or at a _minimum_ to default it off when terrain is present, so that our defaults favor correctness over speed.